added linaer stt
